[mob][photos] Refactor
